# fix(Scripts/SunwellPlateau): Add Blink with Nove Arch Mage