@Rorschach91 2025-04-12
> Not sure whats up with codestyle... I'm not seeing whitespaces... Though is in github web interface

Theese
![trailing ws](https://github.com/user-attachments/assets/7954e6b7-f766-4349-b44a-f8835bb33822)
![trailing ws2](https://github.com/user-attachments/assets/3f4821c0-64b6-42a9-92cb-667870f8fca6)