@blinkysc 2025-04-12
> > Not sure whats up with codestyle... I'm not seeing whitespaces... Though is in github web interface
> 
> Theese for example ![trailing ws](https://private-user-images.githubusercontent.com/108557877/433050148-7954e6b7-f766-4349-b44a-f8835bb33822.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDQ0NjY2MDMsIm5iZiI6MTc0NDQ2NjMwMywicGF0aCI6Ii8xMDg1NTc4NzcvNDMzMDUwMTQ4LTc5NTRlNmI3LWY3NjYtNDM0OS1iNDRhLWY4ODM1YmIzMzgyMi5wbmc_WC1BbXotQWxnb3JpdGhtPUFXUzQtSE1BQy1TSEEyNTYmWC1BbXotQ3JlZGVudGlhbD1BS0lBVkNPRFlMU0E1M1BRSzRaQSUyRjIwMjUwNDEyJTJGdXMtZWFzdC0xJTJGczMlMkZhd3M0X3JlcXVlc3QmWC1BbXotRGF0ZT0yMDI1MDQxMlQxMzU4MjNaJlgtQW16LUV4cGlyZXM9MzAwJlgtQW16LVNpZ25hdHVyZT05ODBhYjk1MDJhZWQxMzdmNGUyY2Y0ZjRhNzI0NGEwMzdkOTQ1OTNiMmRjNmFlMGU4ODJhYzg1NjA5NDBjMjQ5JlgtQW16LVNpZ25lZEhlYWRlcnM9aG9zdCJ9.BwoDrmRhaomp4gCFeB_w7fi1jlrSDp93MRK3K2DqpJ4) ![trailing ws2](https://private-user-images.githubusercontent.com/108557877/433050151-3f4821c0-64b6-42a9-92cb-667870f8fca6.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDQ0NjY2MDMsIm5iZiI6MTc0NDQ2NjMwMywicGF0aCI6Ii8xMDg1NTc4NzcvNDMzMDUwMTUxLTNmNDgyMWMwLTY0YjYtNDJhOS05MmNiLTY2Nzg3MGY4ZmNhNi5wbmc_WC1BbXotQWxnb3JpdGhtPUFXUzQtSE1BQy1TSEEyNTYmWC1BbXotQ3JlZGVudGlhbD1BS0lBVkNPRFlMU0E1M1BRSzRaQSUyRjIwMjUwNDEyJTJGdXMtZWFzdC0xJTJGczMlMkZhd3M0X3JlcXVlc3QmWC1BbXotRGF0ZT0yMDI1MDQxMlQxMzU4MjNaJlgtQW16LUV4cGlyZXM9MzAwJlgtQW16LVNpZ25hdHVyZT01NjU2ZGQxMjlmMjI1YWYxNmE2MDA1ZTkyOGMxYjdmNDYxMWQ0ZTczN2IyMzJmZTVhMzVlMjNmMThkYTgzZjdjJlgtQW16LVNpZ25lZEhlYWRlcnM9aG9zdCJ9.wyf1SXxcNo8op-RRd5Q3CkVeOWkJOpDh-BWMHB8WfK4)

Lol always when I make edits through web interface these show up